Refactor AuthService file path for consistency.
