+++ selsolun-front/src/app/components/register/register.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, NgForm, ReactiveFormsModule } from '@angular/forms';
 
 import { RegisterData } from '../../models/register-data.interface';
-import { AuthService } from '../../services/authentication/auth.service';
+import { AuthService } from '../../services/auth.service';
 import { Router } from '@angular/router';
 
 @Component({
